feat(ast): expose literal parsers used within the parser
The literal parsers take the raw literal that is tokenized by the
scanner and returns the appropriate value type to be used with the ast
literals.
